requirements.txt
Django >= 1.2, < 1.3
django-registration == 0.7
akismet == 0.2.0
FeedParser == 4.1
